ci: add migration for storybook
