--- .storybook/.babelrc
@@ -1,15 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-react",
-    "@babel/preset-typescript"
-  ],
-  "plugins": [
-    [
-      "@babel/plugin-transform-typescript",
-      {
-        "allowNamespaces": true
-      }
-    ]
-  ]
-}
