+++ documentation/tools/nrwl-migrations.md
@@ -1,0 +1,3 @@
+# NRWL Migrations
+
+For local schematics, we'll have to manually add `migrations.json`, also need to update `package.json` to point to `migrations.json`. This has already been added
