+++ libs/test/src/Test.stories.tsx
@@ -1,0 +1,4 @@
+import { Button } from 'antd'
+import React from 'react'
+
+export const Primary = () => <Button>Click</Button>
