+++ libs/tools/plugins/codelab/migrations.json
@@ -2,8 +2,13 @@
   "schematics": {
     "update-0.0.2": {
       "version": "0.0.2",
-      "description": "update-0.0.2",
-      "factory": "./src/migrations/update-0.0.2/update-0.0.2"
+      "description": "update-0-0-2",
+      "factory": "./src/migrations/update-0-0-2/update-0-0-2"
+    },
+    "update-10.3.1": {
+      "version": "10.3.1",
+      "description": "update-10-3-1",
+      "factory": "./src/migrations/update-10-3-1/update-10-3-1"
     }
   },
   "packageJsonUpdates": {
