+++ libs/tools/plugins/codelab/package.json
@@ -1,7 +1,11 @@
 {
   "name": "@codelab/schematics",
-  "version": "0.0.1",
+  "version": "10.3.1",
   "main": "src/index.js",
   "schematics": "./collection.json",
-  "builders": "./builders.json"
+  "builders": "./builders.json",
+  "ng-update": {
+    "requirements": {},
+    "migrations": "./migrations.json"
+  }
 }
