+++ libs/tools/plugins/codelab/src/schematics/library/react/files/.storybook/main.js.template
@@ -1,0 +1,7 @@
+const rootMain = require('<%= offsetFromRoot %>../.storybook/main');
+
+// Use the following syntax to add addons!
+// rootMain.addons.push('');
+rootMain.stories.push(...['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'])
+
+module.exports = rootMain;
