+++ libs/tools/plugins/codelab/src/schematics/library/react/files/.storybook/preview.js.template
@@ -1,0 +1,5 @@
+import { withKnobs } from '@storybook/addon-knobs'
+import { addDecorator } from '@storybook/react'
+import '<%= offsetFromRoot %>/.storybook/preview'
+
+addDecorator(withKnobs)
