+++ libs/tools/plugins/codelab/src/schematics/library/react/files/.storybook/webpack.config.js.template
@@ -1,0 +1,13 @@
+const rootWebpackConfig = require('<%= offsetFromRoot %>/.storybook/webpack.config')
+
+module.exports = async ({ config, mode }) => {
+  // eslint-disable-next-line no-param-reassign
+  config = await rootWebpackConfig({ config, mode })
+
+  config.module.rules.push({
+    test: /\.(ts|tsx)$/,
+    loader: require.resolve('babel-loader'),
+  })
+
+  return config
+}
