--- libs/ui/antd/.storybook/addons.js
@@ -1,1 +1,0 @@
-import '../../../../.storybook/addons'
