--- libs/ui/antd/.storybook/config.js
@@ -1,6 +1,0 @@
-import { withKnobs } from '@storybook/addon-knobs'
-import { addDecorator, configure } from '@storybook/react'
-import '../../../../.storybook/config'
-
-addDecorator(withKnobs)
-configure(require.context('../src', true, /\.stories\.(j|t)sx?$/), module)
