+++ libs/ui/antd/.storybook/main.js
@@ -1,4 +1,4 @@
-const rootMain = require('../../../.storybook/main')
+const rootMain = require('../../../../.storybook/main')
 
 // Use the following syntax to add addons!
 // rootMain.addons.push('');
