+++ libs/ui/antd/.storybook/preview.js
@@ -1,0 +1,5 @@
+import { withKnobs } from '@storybook/addon-knobs'
+import { addDecorator } from '@storybook/react'
+import '../../../../.storybook/preview'
+
+addDecorator(withKnobs)
