+++ migrations.json
@@ -90,6 +90,13 @@
       "factory": "./src/migrations/update-10-3-0/update-10-3-0",
       "package": "@nrwl/storybook",
       "name": "update-10.3.1"
+    },
+    {
+      "version": "10.3.1",
+      "description": "Upgrade Storybook config",
+      "factory": "./src/migrations/update-10-3-1/update-10-3-1",
+      "package": "@codelab/schematics",
+      "name": "update-10.3.1"
     }
   ]
 }
